backend/whisper/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
loguru>=0.7.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.35.0
scipy>=1.11.3
numpy>=1.24.0
soundfile>=0.12.1 